+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Changes:
 
 \-
 
+## v2023-01-31-1
+
+Changes:
+
+- query params
+
 ## v2023-01-03-1
 
 Changes: